chat/requirements.txt
django==4.2
channels==4.0.0
daphne==4.0.0
redis==4.5.4
djangorestframework==3.14.0
psycopg2-binary==2.9.6